requirements.txt
Flask
Werkzeug
tensorflow
numpy
pillow
reportlab
pdfkit